pyproject.toml
[tool.poetry]
name = "image-caption"
version = "0.1.0"
description = ""
authors = ["TobiFank <tobias7402@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
pillow = "^10.4.0"
torch = "^2.4.1"
transformers = "^4.45.1"
openai = "^1.50.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
